Fix package structure and response writer
